almost done really

+++ src/main/java/net/softsociety/aimori/domain/FacilitiesValuation.java
@@ -12,6 +12,6 @@ public class FacilitiesValuation {
 	int 	facilitiesNumber; 			// 시설 번호
 	String 	memberId; 					// 리뷰 작성 회원 아이디
 	String 	facilitiesReview; 			// 시설 리뷰 내용
-	int 	facilitiesStar; 			// 시설 평가(별점)
+	double 	facilitiesStar; 			// 시설 평가(별점)
 	String 	facilitiesReviewDate;	 	// 시설 리뷰 작성 일자
 	}
